+++ flaskblog/models.py
@@ -1,0 +1,27 @@
+from datetime import datetime
+from flaskblog import db
+
+#DATABASE MODELS
+class User(db.Model):
+    id=db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(20), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    image_file = db.Column(db.String(20), nullable=False, default='default.jpg')
+    password = db.Column(db.String(60), nullable=False) #passwords will be hashed!
+    posts = db.relationship('Post', backref='author', lazy=True)
+
+    def __repr__(self):
+        return f"User('{self.username}'), '{self.email}', '{self.email}', '{self.image_file}'"
+
+class Post(db.Model):
+    id=db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(100), nullable=True)
+    date_posted = db.Column(db.DateTime, default=datetime.utcnow)##remove nullable! due to an error
+    ##datetime.utcnow is a function!
+    ##datetime.utcnow() return of a function!
+    content = db.Column(db.Text, nullable=False)
+    user_id=db.Column(db.Integer, db.ForeignKey('user.id'))##remove nullable! due to an error
+#error was-> TypeError: Additional arguments should be named <dialectname>_<argument>, got 'nullabe'
+
+    def __repr__(self):
+        return f"Post('{self.title}', '{self.date_posted}')"
