+++ flaskblog/static/favicon.ico
