+++ requirements.txt
@@ -1,7 +1,17 @@
 blinker==1.6.2
 click==8.1.7
+dnspython==2.4.2
+email-validator==2.0.0.post2
 Flask==2.3.3
+Flask-SQLAlchemy==3.1.1
+Flask-WTF==1.1.1
+idna==3.4
+importlib-metadata==6.8.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.3
+SQLAlchemy==2.0.21
+typing_extensions==4.8.0
 Werkzeug==2.3.7
+WTForms==3.0.1
+zipp==3.16.2
